[FIX+IMP] account_invoice_inter_company: Remove VAT + useless records
Since https://github.com/odoo/odoo/pull/68253, we can't put invalid
VAT numbers anymore, but they are not needed for the purpose of the
module, so we remove the field assignation.

Visiting this, the fiscal positions have been detected as not used, so
we remove them for cleaning a bit the module test.