[18.0][MIG] product_tax_multicompany_default: migration to 18.0
This PR is a continuation of PR https://github.com/OCA/multi-company/pull/811.

This PR is for the migration of the product_tax_multicompany_default module, which is used to propagate taxes across multiple companies.

This commit makes changes so that it is not duplicated in the Sales Taxes form view, and removes the functionality when creating a new product in all companies, since Odoo 18 already does this.

@moduon MT-11041