[FIX] purchase_sale_inter_company: avoid multicompany issues when the sale order is created and confirmed
Install the purchase_sale_stock_inter_company module

Configuration in the companies:
- Sync pickings
- Auto-validate the sale order

Enable Dropshipping and Multi-Step Routes in the inventory settings.
In the Dropship route, set Company A.
In the Buy route, set Company B.

Create a product and, in Company A, enable only the Dropshipping route.
Set the supplier info with the partner of Company B, but do not set another supplier info for Company B (because this product should not automatically generate a purchase in Company B).

Create a sale order with any customer, confirm it, and go to the generated purchase order.
Try to confirm the purchase order, an error will be displayed:

“There is no matching vendor price to generate the purchase order for this product.”

This happens because the sale order in Company B is confirmed automatically, and the code executes with sudo, meaning record rules are not applied. As a result, the Dropship route is considered, although in this case it should be ignored.

After this commit, record rules are correctly applied, and routes are taken into account according to the record rules for the current company.